+++ guillotina/__init__.py
@@ -9,8 +9,6 @@ from guillotina._cache import FACTORY_CACHE  # noqa
 from guillotina._cache import PERMISSIONS_CACHE  # noqa
 from guillotina._cache import SCHEMA_CACHE  # noqa
 from guillotina._settings import app_settings  # noqa
-from guillotina._settings import request_var  # noqa
-from guillotina._settings import txn_var  # noqa
 from guillotina.i18n import default_message_factory as _  # noqa
 from zope.interface import Interface  # noqa